extract callback and logic into data container; convert to TS
